refactor: add single axios instance and interceptor for handling token expiry

+++ src/axios/auth.requests.ts
@@ -1,18 +1,18 @@
-import axios from 'axios';
+import api from './axios';
 import { ILogin } from '../interfaces';
 
-const api = axios.create({
-  baseURL: 'http://localhost:3000/api/v1/auth',
-});
 
 export const loginUser = async (newUser: ILogin): Promise<boolean> => {
-  const response = await api.post('/login', newUser, { withCredentials: true });
-  // console.log("Response after login api: ",response.data);
+  const response = await api.post('/auth/login', newUser, { withCredentials: true });
+
+  //set refresh token in localstorage
+  localStorage.setItem("refreshToken", response.data.data)
+
   return response.data.status;
 };
 
 
 export const logoutUser = async (): Promise<void> => {
-    const response = await api.delete('/logout', { withCredentials: true })
+    const response = await api.delete('/auth/logout', { withCredentials: true })
     return response.data;
 }
